Created rough draft of Ephemeral Snake Pattern (5)
A bit buggy - but Pattern5 will randomly create snakes which start at
length 1, grow to a random length in a certain direction, and then
shrink down to nothing. I think it'll look super cool.

Also created an array of the primary LITColors, for situations where we
need to randomly select a primary color.