Added Shift Register Shit
Added beginnings of a shift register library. Archived the code
(although I guess github also does this) in case I want to keep editing
a ton of crap. This is still all on the dev branch anyway, so I am
probably good to edit a lot without fear of losing work.