Initial round of functionality tweaks and overhauls
Went through code and discovered changes I wanted to make to ensure
functionality was in place for hardcore testing of the new
architecture.

Parameters.h - created this to hold relevant max values that can be
seen by all classes. Couldn't include in LIT.h because of recursive
behavior with LED class. Went through code and replaced instances of
numbers (e.g. 100) with these variables (e.g. maxBrightness).

LIT.h - #included a bunch of libraries to slim down #include statements
in thing, pattern, and hb.

LITColor.h - created 2:1:0 colors. Did my best at naming them.

LED - created LayeredLED and AddressedLED, which are both subclasses of
LED. I removed layer as a variable from LED, bringing its size down to
4 bytes instead of 5. The global array of 32 leds is now of type
LayeredLED, which simply contains the layer variable on top of its LED
functionality. I did this to reduce the file size of Thing by 32 bytes.
I made the AddressedLED class in case we need to eliminate the massive
32-led array from the Thing class and instead only allow a thing to own
the LEDs it absolutely needs. This would have the thing owning
AddressedLEDs, which have an address as an extra member variable.

Pattern updating - included and tested functionality for mixing colors
and brightnesses between LEDs, as well as the layering functionality.
All is well and ready to go.