Created LinkedList class
Made a linked list class as an alternative to our current methods of
creating LED arrays in Thing().

Currently it's able to append entries to the end, remove entries,
return a given entry, and calculate list length.

More rigorous error checking needs to be implemented, but basic
functionality exists. A tester function was created to test it and show
it off.