First successes with using a list of steps to update a foo
Was able to have a foo go through two different steps for different
numbers of iterations. Despite the crazy syntax, I confirmed the data
was being manipulated as expected.

Currently the only functions that can be passed to steps are void
functions that take no arguments. Custom functions can be created
within the custom foo and can be added to the steps list as long as
they are cast correctly.