Created listener foo as first attempt at detecting beats
Listener foo shoots a 'moving dot' around the headband if a beat is
detected. In the simulator, pressing the event button simulates a beat
on band 0. Works.

Not sure about the best way to interface with the audio object from
outside of it. But I've got the thing to work.