Created SRColor function in LIT.h
Created this function solely to be able to convert the leds data to
shift register data. Returns a "color" based on the index of the RGB
LED shift register. Since the wiring does not at all correspond 1:1
with the actual LEDs being stored, this is what we have to do.

So far, the function is fairly efficient. May be improved with some
better math.