Create SRTrio class to deal with writing to shift registers
Originally I had just used the SRColor function to create a “color” to
hand to shift registers. I needed to do this because the LEDs are all
wired out of order, so I had created a function to order them.

As I was trying to implement brightness controls, I realized something
was amiss - that brightness was not being recognized properly at the SR
level. Then I realized it was because I was passing it a color, which
expands its r,g,b values to the maximum ratio it can handle after being
set.

Since, under our scheme, a color is merely a ratio between r,g,b
values, I needed to create a different data structure that more truly
represents what it is doing - namely, SRTrio, or a trio of values that
I can pass to a shift register to fool it into thinking it’s an LED.

Now that SRTrio doesn’t expand the three values or treat them like a
ratio, I can pass in a brightness adjusted value and it will all work
out well.

For now, at least. I’m assuming this is still sort of a patch.