Create battery monitoring foo
DoubleMonitor monitors both the battery voltage and the 3V signal. It
outputs these values as binary numbers so I can read them without being
plugged into the serial port.

This is designed to monitor the arduino readings vs multimeter
readings. The resulting data should give us a good idea of the range of
VBAT readings, and consequently can return battery percentage to the
user and potentially power off the arduino if the reading gets too low.