fixed it
two problems from before: I was using = operator to assign bandColor
instead of the setColor function. I was also continuing the for loop
where I check every band for a beat instead of breaking as soon as I
found one had a beat. This caused some problems because I was using the
addLEDs function over and over in a short period of time. Beat
Detection still sucks in my opinion. Many false positives and not
detecting when it should.